src/requirements.txt
django >= 2.1.6, < 3.0 
djangorestframework
django-cors-headers
channels
channels_redis
asgi_redis
pypiwin32